refactor(eslint-effect): rename no-switch-on-tag to no-switch-statement
The rule now bans ALL switch statements, not just switches on _tag,
so the name should reflect its actual behavior.
