+++ packages/eslint-effect/src/rules/no-switch-statement.js
