+++ packages/eslint-effect/test/no-switch-statement.test.ts
@@ -7,7 +7,7 @@ type Event =
   | { readonly type: 'Updated'; readonly name: string };
 const event: Event = { type: 'Created', name: 'test' } as Event;
 
-// eslint-disable-next-line effect/no-switch-on-tag, effect/no-direct-tag-access -- Testing switch on _tag
+// eslint-disable-next-line effect/no-switch-statement, effect/no-direct-tag-access -- Testing switch on _tag
 switch (either._tag) {
   case 'Right':
     break;
@@ -15,7 +15,7 @@ switch (either._tag) {
     break;
 }
 
-// eslint-disable-next-line effect/no-switch-on-tag -- Testing switch on type discriminator
+// eslint-disable-next-line effect/no-switch-statement -- Testing switch on type discriminator
 switch (event.type) {
   case 'Created':
     break;
@@ -25,7 +25,7 @@ switch (event.type) {
     break;
 }
 
-// eslint-disable-next-line effect/no-switch-on-tag -- Testing switch on any value
+// eslint-disable-next-line effect/no-switch-statement -- Testing switch on any value
 switch ('foo' as string) {
   case 'foo':
     break;
@@ -35,7 +35,7 @@ switch ('foo' as string) {
     break;
 }
 
-// eslint-disable-next-line effect/no-switch-on-tag -- Testing switch on number
+// eslint-disable-next-line effect/no-switch-statement -- Testing switch on number
 switch (42 as number) {
   case 42:
     break;
